Walletd RPC : Changed input key block_height to block_index in commands get_transaction and get_transactions
Changed block_height to block_index because the value being returned is in fact the index of the block and not the height of the block. The height of a block is equal to the index of the block + 1.

Reason for change : Previous name of block_height was incorrect and misleading. The correct term is block index.