SimpleWallet Commands : Fixed block index vs block height for command height
Also made changes to function updateBlockchainStatus() in NodeRpcProxy/NodeRpcProxy.cpp because we adjusted the value of rsp.block_header.height in the Daemon JSON RPC command get_last_block_header to return the block height and not the block index. But updateBlockchainStatus() uses the block index so we have to subtract 1 from rsp.block_header.height to give the block index. Also m_nodeHeight acutally stores the top local block index and should be renamed to m_topBlockIndex. Also getLastLocalBlockHeight() and getLocalBlockCount() acutally return the block index and not the block height and should be renamed to getTopLocalBlockIndex().

Reason for changes : Trying to fix the mix up between block index and block height in the code and also the bug that was introduced in commit dd183312dd323b0318c07709f0bbc723a655fd1f when command get_last_block_header was changed to return the block height and not the block index