SimpleWallet : Fixed syncrhonization height
When SimpleWallet is opened and shows that the wallet is synchronizing, the block height was showing the block index rather than the block height. This commit changes the code to now show the block height.

Also, WalletLegacyTransaction.blockHeight should be renamed to blockIndex and getLastLocalBlockHeight() in NodeRpcProxy should be renamed to getLastLocalBlockIndex().

Reason for changes : To clear up the confusion between block index and block height in the code.