Walletd : Fixed synchronization height shown when starting Walletd
When Walletd is opened and shows that the wallet is synchronizing, the block height was showing the block index rather than the height. This commit changes the code to now show the block height.

Reason for changes : To clear up the confusion between block index and block height in the code.