Blockchain : Rename BlockEntry.height to BlockEntry.block_index
Reason for changes : In function Blockchain::pushBlock(), block.height = static_cast<uint32_t>(m_blocks.size());. This occurs before the block is added to the blockchain. Because  block.height is 1 less than the blockchain height after the block is added, we can conclude that block.height acutally stores the block's index and not the block's height. Therefore, we renamed block.height to block.block_index. This change is to help clear of the confusion in the codde between the terms block height and block index.