Checkpoints : Changed CheckpointData.height to CheckpointData.blockIndex
Reason for change : In CryptoNoteConfig.h, the current checkpoints listed are in the form {blockIndex, blockHash}. Also, in CryptoNoteCore/Blockchain.cpp function pushBlock() a line is called m_checkpoints.check_block(getCurrentBlockchainHeight(), blockHash). Here the getCurrentBlockchainHeight() returns the index of the block being added because at this point the block has not been added to the blockchain yet. Therefore, the checkpoints that we add need to be in the form of {blockIndex, blockHash}. This is to help clarify the confusion in the code between block index and block height.

In CryptoNoteCore/Checkpoints.cpp function check_block