Unit Tests : Fixed WalletLegacyApi.checkChange test from failed
Reason for changes : Originally the value fro banknote was equal to the transaction fee which is 10000000. Therefore, the transaction would fail because the amount sent plus the fee was greater than the banknote variable. Increasing the banknote variable to more than the transaction fee plus the sentAmount allows this test to pass.