Core Tests : Fixed multiple core tests to make them runnable
Reason for changes : The transaction amounts that were being created in the core test code were too large because Cash2 has a block reward of 0.894069671 or less, so the accounts did not have enough money after mining on a few blocks to send transactions with such high amounts. This caused the tests to fail.