Daemon JSON RPC : Added command validate_address
Checks whether a Cash2 address is valid or not. Also made a change to CMakeLists.txt in the "tests" folder to fix compilation failure on Windows computers.

Reason for change :  this feature was requested by the exchange SouthXchange

This is a copy of the commit b43acfd2c6317f511c17ba02ee0b626d6c963261 from the southXchange branch