Walletd RPC : Added command validate_address
Checks whether a Cash2 address is valid or not.

Reason for change : SouthXchange has requested this RPC command to be added to the Walletd program instead of the daemon to make integration and future upgrades easier.

This is a copy of the commit 6998417bef73b2f887fc1f802544ac81ab21b77e from the southXchange branch