P2p : Removed unused files from P2p folder
Removed unused files from P2p folder and from P2p test folders. If these files are needed in the future, they can be put back in, but right now there does not seem to be a need for these files.

Reason for changes : to only keep useful code in the repository