SimpleWallet : Using size_t instead of CryptoNote::TransactionId for the transaction index for Wallet Legacy
For Wallet Legacy, the transaction index type used to by TransactionId which was an alias for size_t. This code removes the type TransactionId and uses size_t instead.

Reason for changes : the name TransactionId was confusing because it wasn't really an id, it was just an index. Renaming TransactionId to TransactionIndex caused a naming conflict with the Wallet folder. So we decided to just remove the TransactionId type and use size_t instead