SimpleWallet : Using size_t instead of CryptoNote::TransferId for the transfer index for Wallet Legacy
For Wallet Legacy, the transfer index type used to be TransferId which was an alias for size_t. This code removes the type TransferId and uses size_t instead. Also renamed the variable names to indicate that it is transfer index rather than a transfer id.

Reason for changes : TransferId was misnamed because it wasn't really an id but more of an index. Using size_t instead simplifies the code.