Walletd : Removed class PaymentGateService and move all code directly into the main() function
All code in class PaymentGateService is move directly into the main() function. Removed files PaymentGateService.h and PaymentGateService.cpp.

Reason for changes : To reduce the number of classes and files and to try and keep all functionality inside the main() function. This will hopefully help make the code easier to read and understand for newcomers.