Walletd : Added an event lock to function saveWallet() and rearranged some minor logging information
Added an event lock to function saveWallet() as a precaution. I am not 100% sure if this is needed, but I will add it here to be safe. This can be removed in the future if needed. Also moved the logging information inside the function where the file saving is actually occurring.

Reason for changes : added an event lock as a precaution