Walletd : Refactored class WalletGreen and fixed compilation error on Linux
Refactored class WalletGreen. Still not finished because there is a lot of complicated code. In the future, we need to simplify the code in this class. I also fixed 3 lines of code that was causing compliation error. 2 lines in functions createAddress() in file WalletHelper.cpp and 1 line in function TEST_F(WalletHelperTest_createAddress, correctPublicKey) in file TestWalletHelper.cpp. For some reason, compilation will succeed on Windows but fail on Linux computers.

Reason for changes : to better organize the WalletGreen class to make the code easier to read and understand and to fix compilation error on Linux computers.