Walletd : RPC commands createAddress() and createAddress(spendPrivateKey) no longer return the spend private key of the address created
Reason for changes : Previously in commit 208d1f3, a change was made where the createAddress() RPC commands would return the spend private key of the newly created address. This was put in place to make it easier to the caller to keep record of the new spend private key. I was asked by the SouthXchange team to remove this feature as they felt that passing private keys outside of the wallet file is a security risk.

This commit is a close duplicate of commit 00e2a7f8b2b68397f4e694e29710f6ccb6a4c9ff from the exchanges branch