Compilation : Compile Cash2 software on ARM64 architecture
SouthXchange has requested the Cash2 daemon and CLI wallet be able to run on ARM64 servers because they are switching to using ARM64 servers. CMakeLists.txt was modified to include detecting ARM architecture, and the slowhash function was modified to work on ARM architecture. The new code for the slowhash function was borrowed from The Karbo developers. This code was tested and will still compile on Linux and Windows in addition to being able to compile on a ARM64 server at Amazon Web Services running Ubuntu Server 18.04 LTS.